move category and expense forms default values to constants.ts
